requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
audioop-lts==0.2.2
audioread==3.1.0
banal==1.0.6
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
cycler==0.12.1
dataset==1.6.2
datasets==4.5.0
decorator==5.2.1
dill==0.4.0
filelock==3.24.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
intel-cmplr-lib-ur==2025.3.2
intel-openmp==2025.3.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mkl==2025.3.1
mkl-service==2.6.1
mkl_fft==2.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
narwhals==2.16.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
onemkl-license==2025.3.1
packaging==26.0
pandas==2.3.3
pillow==12.1.1
platformdirs==4.9.2
pooch==1.9.0
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==1.4.54
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
streamlit==1.54.0
sympy==1.14.0
tbb==2022.3.1
tcmlib==1.4.1
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.3
transformers==5.2.0
triton==3.6.0
typer==0.24.1
typer-slim==0.24.0
typing_extensions==4.15.0
tzdata==2025.3
umf==1.0.3
urllib3==2.6.3
watchdog==6.0.0
xxhash==3.6.0
yacs==0.1.8
yarl==1.22.0
gdown
